requirements.txt
# Fingerprint Recognition Framework Requirements
# ================================================

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Image processing
opencv-python>=4.5.0
scikit-image>=0.18.0
Pillow>=8.0.0

# Machine learning
scikit-learn>=0.24.0

# Deep learning (optional - for CNN models)
torch>=1.9.0
torchvision>=0.10.0

# Configuration management
PyYAML>=5.4.0

# Visualization (optional - for plots)
matplotlib>=3.4.0
seaborn>=0.11.0

# Streamlit UI
streamlit>=1.28.0

# Data handling
tqdm>=4.60.0

# Development tools (optional)
pytest>=6.2.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.900

# Jupyter notebooks (optional)
jupyter>=1.0.0
ipykernel>=6.0.0